reorgize planner components
